+++ components/leaderboard/LeaderboardTabs.tsx
@@ -40,12 +40,12 @@ export const LeaderboardTabs = ({ activeTab, onTabChange }: LeaderboardTabsProps
             paddingHorizontal: 20,
             paddingVertical: 8,
             backgroundColor: activeTab === 0 ? themeColor : 'transparent',
-            color: activeTab === 0 ? '#222' : '#FFF',
-            borderRadius: 6,
+            color: activeTab === 0 ? '#222' : '#222',
+            borderRadius: 20,
             fontSize: 20,
             fontWeight: 'bold',
             fontFamily: 'magistral',
-            shadowColor: activeTab === 0 ? '#F23B3B' : 'transparent',
+            shadowColor: activeTab === 0 ? themeColor : 'transparent',
             shadowOffset: { width: 0, height: 0 },
             shadowOpacity: 0.8,
             shadowRadius: 12,
@@ -60,12 +60,12 @@ export const LeaderboardTabs = ({ activeTab, onTabChange }: LeaderboardTabsProps
             paddingHorizontal: 20,
             paddingVertical: 8,
             backgroundColor: activeTab === 1 ? themeColor : 'transparent',
-            color: activeTab === 1 ? '#222' : '#FFF',
-            borderRadius: 6,
+            color: activeTab === 1 ? '#222' : '#222',
+            borderRadius: 20,
             fontSize: 20,
             fontWeight: 'bold',
             fontFamily: 'magistral',
-            shadowColor: activeTab === 1 ? '#F23B3B' : 'transparent',
+            shadowColor: activeTab === 1 ? themeColor : 'transparent',
             shadowOffset: { width: 0, height: 0 },
             shadowOpacity: 0.8,
             shadowRadius: 12,
